+++ CHANGELOG.md
@@ -4,8 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: `IconBadge`: Add `border` and `backgroundColor` props style.
-- Fix: `ButtonFix`: Fix min width constraints
+## [5.1.0] - 2021-10-15
+
+Feature:
+
+- `IconBadge`: Add `border` and `backgroundColor` props style.
+
+Fix:
+
+- `Button`: Fix min width constraints.
 
 ## [5.0.2] - 2021-10-13